# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Parking System API",
    "version": "1.0.0",
...
    "contact": {
      "name": "API Support",
      "email": "support@smartparking.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://smart-parking-backend-production-5449.up.railway.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Floor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "name": {
            "type": "string",
            "example": "Ground Floor"
          },
          "level": {
            "type": "integer",
            "example": 0,
            "nullable": true
          },
          "parkingLot": {
            "$ref": "#/components/schemas/ParkingLot"
          },
          "parkingSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParkingSlot"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ParkingSlot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "name": {
            "type": "string",
            "example": "A-001"
          },
          "isReservable": {
            "type": "boolean",
            "example": false
          },
          "floor": {
            "$ref": "#/components/schemas/Floor"
          },
          "node": {
            "$ref": "#/components/schemas/Node"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/floors": {
      "get": {
        "summary": "Get all floors for current admin",
        "tags": [
          "Floors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All floors retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "All floors retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Floor"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
...
    "/api/floors/{id}": {
      "get": {
        "summary": "Get floor by ID",
        "tags": [
          "Floors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            },
            "description": "Floor ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Floor retrieved successfully"
          },
          "404": {
            "description": "Floor not found"
          }
        }
      },
      "put": {
        "summary": "Update floor",
        "tags": [
          "Floors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            },
            "description": "Floor ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "level": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Floor updated successfully"
          },
          "404": {
            "description": "Floor not found"
          }
        }
      },
      "delete": {
        "summary": "Delete floor",
        "tags": [
          "Floors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            },
            "description": "Floor ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Floor deleted successfully"
          },
          "400": {
            "description": "Cannot delete floor with existing parking slots"
          },
          "404": {
            "description": "Floor not found"
          }
        }
      }
    },
...
    "/api/gateways/nodes": {
      "post": {
        "summary": "Create node under gateway (Admin only)",
        "tags": [
          "Nodes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "gatewayId",
                  "chirpstackDeviceId",
                  "name"
                ],
                "properties": {
                  "gatewayId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  },
                  "chirpstackDeviceId": {
                    "type": "string",
                    "example": "sensor_001"
                  },
                  "name": {
                    "type": "string",
                    "example": "Parking Slot A1 Sensor"
                  },
                  "description": {
                    "type": "string",
                    "example": "Ultrasonic sensor for slot A1"
                  },
                  "latitude": {
                    "type": "number",
                    "example": 40.7128
                  },
                  "longitude": {
                    "type": "number",
                    "example": -74.006
                  },
                  "metadata": {
                    "type": "object",
                    "example": {
                      "sensorType": "ultrasonic",
                      "range": "4m"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Node created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
...
    "/api/gateways": {
      "post": {
        "summary": "Create a new gateway (Super Admin only)",
        "tags": [
          "Gateways - Super Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "chirpstackGatewayId",
                  "name"
                ],
                "properties": {
                  "chirpstackGatewayId": {
                    "type": "string",
                    "example": "gw_001"
                  },
                  "name": {
                    "type": "string",
                    "example": "Main Building Gateway"
                  },
                  "description": {
                    "type": "string",
                    "example": "Primary gateway for main building parking"
                  },
                  "location": {
                    "type": "string",
                    "example": "Building A, Floor 1"
                  },
                  "latitude": {
                    "type": "number",
                    "example": 40.7128
                  },
                  "longitude": {
                    "type": "number",
                    "example": -74.006
                  },
                  "metadata": {
                    "type": "object",
                    "example": {
                      "model": "RAK7249",
                      "version": "1.0.3"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Gateway created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Access denied"
          }
        }
      },
      "get": {
        "summary": "Get all gateways (Super Admin only)",
        "tags": [
          "Gateways - Super Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "includeInactive",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "default": "false"
            },
            "description": "Include inactive gateways"
          }
        ],
        "responses": {
          "200": {
            "description": "Gateways retrieved successfully"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/api/gateways/{id}": {
      "get": {
        "summary": "Get gateway by ID with nodes (Super Admin only)",
        "tags": [
          "Gateways - Super Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Gateway ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Gateway retrieved successfully"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Gateway not found"
          }
        }
      },
      "put": {
        "summary": "Update gateway (Super Admin only)",
        "tags": [
          "Gateways - Super Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Gateway ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Gateway Name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated description"
                  },
                  "location": {
                    "type": "string",
                    "example": "Building B, Floor 2"
                  },
                  "latitude": {
                    "type": "number",
                    "example": 40.7589
                  },
                  "longitude": {
                    "type": "number",
                    "example": -73.9851
                  },
                  "isActive": {
                    "type": "boolean",
                    "example": true
                  },
                  "metadata": {
                    "type": "object",
                    "example": {
                      "firmware": "2.1.0"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gateway updated successfully"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Gateway not found"
          }
        }
      },
      "delete": {
        "summary": "Delete gateway (Super Admin only)",
        "tags": [
          "Gateways - Super Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Gateway ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Gateway deleted successfully"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Gateway not found"
          }
        }
      }
    },
...
    "/api/gateways/{id}/nodes": {
      "get": {
        "summary": "Get nodes for a gateway",
        "tags": [
          "Nodes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Gateway ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Gateway nodes retrieved successfully"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Gateway not found"
          }
        }
      }
    },
...
    "/api/parking-lots": {
      "get": {
        "summary": "Get all parking lots for the authenticated admin",
        "tags": [
          "Parking Lots"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Parking lots retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ParkingLot"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new parking lot",
        "tags": [
          "Parking Lots"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "address"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Main Building Parking"
                  },
                  "address": {
                    "type": "string",
                    "example": "123 Main Street, City"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Parking lot created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
...
}